# product-api/swagger.yaml
basePath: /
consumes:
    - application/json
definitions:
    Product:
        description: Product defines the structure for an API product
        properties:
            description:
                type: string
                x-go-name: Description
            id:
                description: the id for the product
                format: int64
                minimum: 1
                type: integer
                x-go-name: ID
            name:
                type: string
                x-go-name: Name
            price:
                format: float
                type: number
                x-go-name: Price
            sku:
                type: string
                x-go-name: SKU
        required:
            - id
        type: object
        x-go-package: github.com/MousaZa/product-services/product-api/data
info:
    description: Documentation for Product API
    title: of Product API
    version: 1.0.0
paths:
    /products:
        get:
            description: Returns a list of products
            operationId: listProducts
            responses:
                "200":
                    $ref: '#/responses/productsResponse'
            tags:
                - products
        post:
            operationId: addProduct
            parameters:
                - description: The product to add to the database
                  in: body
                  name: product
                  required: true
                  schema:
                    $ref: '#/definitions/Product'
                  x-go-name: Product
            responses:
                "201":
                    $ref: '#/responses/noContent'
            tags:
                - products
    /products/{id}:
        delete:
            operationId: deleteProduct
            parameters:
                - description: the id of the product to delete from the database
                  format: int64
                  in: path
                  name: id
                  required: true
                  type: integer
                  x-go-name: ID
            responses:
                "201":
                    $ref: '#/responses/noContent'
            tags:
                - products
        get:
            description: Returns a product
            operationId: listSingleProduct
            parameters:
                - description: the id of the product to delete from the database
                  format: int64
                  in: path
                  name: id
                  required: true
                  type: integer
                  x-go-name: ID
            responses:
                "200":
                    $ref: '#/responses/productResponse'
            tags:
                - products
        put:
            operationId: updateProduct
            parameters:
                - description: the id of the product to delete from the database
                  format: int64
                  in: path
                  name: id
                  required: true
                  type: integer
                  x-go-name: ID
                - description: The product to add to the database
                  in: body
                  name: product
                  required: true
                  schema:
                    $ref: '#/definitions/Product'
                  x-go-name: Product
            responses:
                "201":
                    $ref: '#/responses/noContent'
            tags:
                - products
produces:
    - application/json
responses:
    noContent:
        description: ""
    productResponse:
        description: A product returns in the response
        schema:
            $ref: '#/definitions/Product'
    productsResponse:
        description: A list of products returns in the response
        schema:
            items:
                $ref: '#/definitions/Product'
            type: array
schemes:
    - http
swagger: "2.0"